requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astroid==2.4.2
attrs==20.2.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
gym==0.17.2
idna==2.10
imageio==2.9.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
kornia==0.2.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numpy==1.19.1
oauthlib==3.1.0
open3d==0.10.0.0
opencv-python==4.4.0.42
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.9.5
pycparser==2.20
pygame==2.0.0.dev6
pyglet==1.5.0
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.9.3
toml==0.10.1
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.48.2
traitlets== 4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
